+++ client/src/planning/PlannerResponseHandler.ts
@@ -6,7 +6,7 @@
 
 import { Plan } from "../../../common/src/Plan";
 
-export interface PlanningHandler {
+export interface PlannerResponseHandler {
     handleOutput(outputText: string): void;
     handleSuccess(stdout: string, plans: Plan[]): void;
     handleError(error: Error, stderr: string): void;